Added more functions
generateRefreshToken() - Creates a Refresh Token that only contains the User ID and it expires in one day. This is different from the Access Token because the Access Token contains ALL User information (including User ID) but expires in 15 mins during user session.
authUser() - Checks if Access Token is valid (Access Token is valid when it matches JWT_SECRET and is not expired). If Access Token is valid, user information is extracted and sent back to the client. The client will set this data using a useState() (An example exists in homepage.js, found in React frontend).
refreshAcesssToken() - If authUser() receives an invalid Access Token, this function is called. It will check the Refresh Token and see if it exists in the cookies. If it does and is validated, the User Id will be extracted from the Refresh Token. The User Id will be used to search through the database and if the user is found, it will take all the user information and package it into a new Access Token. This new Access Token will be send back to the client and will be used to extract user information.